// e-commerce/ecommerce-angular/src/app/model/product.ts
export class Product{
  color: any;
  cpu: any;
  gpu: any;
  hdd: any;
  operatingSystem: any;
  productBrand: any;
  productCode: any;
  productId: any;
  productImage: any;
  productPrice: any;
  productScore: any;
  productTitle: any;
  ram: any;
  screenSize: any;
  usageType: any;
  weight: any;
}